==> CDD_training_aids/New_Stuff/FootS/2_Normalize.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_Analysis/New_Stuff/FootS/FootS.rda")
FootS1m$Area<-as.numeric(FootS1m$Area)
bromo <- FootS1m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_FootS <- FootS1m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_FootS)) {
  split_FootS[[j]]$Area <- split_FootS[[j]]$Area / unlist(bromo[j, "Area"])
}
FootS1m <- do.call("rbind", split_FootS)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/FootS/FootSn.rda", FootS1m)
```
